+++ ask-sdk-core/package.json
@@ -1,11 +1,11 @@
 {
   "name": "ask-sdk-core",
-  "version": "2.0.1",
+  "version": "2.0.2-SNAPSHOT",
   "description": "Core package for Alexa Skills Kit SDK",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "default": "gulp"
+    "default": "./node_modules/.bin/gulp"
   },
   "author": "Amazon.com",
   "contributors": [
