+++ ask-sdk-dynamodb-persistence-adapter/tsconfig.json
@@ -1,18 +1,16 @@
 {
-    "compilerOptions": {
-        "target": "es5",
-        "module": "commonjs",
-        "moduleResolution": "node",
-        "outDir" : "dist",
-        "sourceMap" : true,
-        "inlineSources" : false,
-        "alwaysStrict" : true,
-        "removeComments" : false,
-        "declaration" : true,
-        "lib": ["es2017", "dom"],
-        "typeRoots": [
-            "./node_modules/@types"
-        ]
-    },
-    "include" : ["lib/**/*.ts", "index.ts"]
+  "compilerOptions": {
+    "target": "es5",
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "rootDir": "lib",
+    "outDir" : "dist",
+    "sourceMap" : true,
+    "inlineSources" : false,
+    "alwaysStrict" : true,
+    "removeComments" : false,
+    "declaration" : true,
+    "lib": ["es2017", "dom"]
+  },
+  "include": ["lib"]
 }
