+++ ask-sdk/package.json
@@ -1,11 +1,11 @@
 {
   "name": "ask-sdk",
-  "version": "2.0.0",
+  "version": "2.0.1-SNAPSHOT",
   "description": "Standard distribution package for Alexa Skills Kit SDK",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "default": "gulp"
+    "default": "./node_modules/.bin/gulp"
   },
   "author": "Amazon.com",
   "contributors": [
